requirements.txt
numpy==1.14.2
matplotlib==2.1.1
dm_sonnet==1.14
attrdict==2.0.0
scipy==0.18.1
orderedattrdict==1.5
#tensorflow==1.6.0
